+++ packages/axway-cli-oum/src/org/activity.js
@@ -20,7 +20,10 @@ export default {
     ${style.highlight('axway org activity <org>')}
 
   Display organization activity for a specific date range:
-    ${style.highlight('axway org activity <org> --from 2021-04-01 --to 2021-04-30')}`;
+    ${style.highlight('axway org activity <org> --from 2021-04-01 --to 2021-04-30')}
+
+  Display organization activity for the current month:
+    ${style.highlight('axway org activity <org> --month')}`;
 		}
 	},
 	options: {
@@ -32,6 +35,10 @@ export default {
 		'--json': {
 			callback: ({ ctx, value }) => ctx.jsonMode = value,
 			desc: 'Outputs the org activity as JSON'
+		},
+		'--month [mm|yyyy-mm]': {
+			desc: 'A month date range; overrides --to and --from',
+			redact: false
 		},
 		'--to [yyyy-mm-dd]': {
 			desc: 'The end date',